move strings to separate file, change name of Register screen
