+++ CookBook/app/(tabs)/_layout.tsx
@@ -2,8 +2,8 @@ import React from 'react'
 import AntDesign from '@expo/vector-icons/AntDesign'
 import { Tabs } from 'expo-router'
 import FontAwesome6 from '@expo/vector-icons/FontAwesome6'
-import { COLORS, ROUTES, ICONS, UI_LABELS } from '@/constants/Constants'
-
+import { COLORS, ROUTES, ICONS } from '@/constants/Constants'
+import { UI_LABELS } from '@/constants/Strings'
 function TabBarIcon(props: {
   name: React.ComponentProps<typeof AntDesign>['name']
   color: string
