+++ CookBook/constants/Constants.ts
@@ -22,15 +22,6 @@ export const ROUTES = {
   TABS_MAIN: '(tabs)',
 } as const
 
-export const ALERTS = {
-  OOPS: 'Oops!',
-  LOGIN_FAILED: 'Login error:',
-  REGISTRATION_FAILED: 'Registration failed',
-  INVALID_CREDENTIALS: 'Invalid email or password',
-  USER_EXISTS: 'User already exists',
-  USER_NOT_FOUND: 'user_not_found',
-} as const
-
 export const FONT_STYLES = {
   WEIGHT: {
     BOLD: 'bold',
@@ -50,16 +41,6 @@ export const LAYOUT = {
 export const ICONS = {
   BOWL_FOOD: 'bowl-food',
   PROFILE: 'profile',
-} as const
-
-export const UI_LABELS = {
-  RECIPES_LIST: 'List of recipes',
-  PROFILE: 'Profile',
-  EMAIL: 'Email',
-  PASSWORD: 'Password',
-  SEARCH: 'Search',
-  ERROR: 'Error',
-  NAME: 'name',
 } as const
 
 export const AUTH = {
