+++ CookBook/constants/Strings.ts
@@ -1,0 +1,17 @@
+export const ALERTS = {
+  OOPS: 'Oops!',
+  LOGIN_FAILED: 'Login error:',
+  REGISTRATION_FAILED: 'Registration failed',
+  INVALID_CREDENTIALS: 'Invalid email or password',
+  USER_EXISTS: 'User already exists',
+  USER_NOT_FOUND: 'user_not_found',
+} as const
+export const UI_LABELS = {
+  RECIPES_LIST: 'List of recipes',
+  PROFILE: 'Profile',
+  EMAIL: 'Email',
+  PASSWORD: 'Password',
+  SEARCH: 'Search',
+  ERROR: 'Error',
+  NAME: 'name',
+} as const
