+++ CookBook/package.json
@@ -19,8 +19,8 @@
     "@react-navigation/native": "^7.1.6",
     "@tanstack/react-query": "^5.85.5",
     "eas-cli": "^16.18.0",
-    "@tanstack/react-query": "^5.85.5",
     "expo": "~53.0.22",
+    "expo-dev-client": "~5.2.4",
     "expo-font": "~13.3.2",
     "expo-linking": "~7.1.7",
     "expo-router": "~5.1.5",
@@ -39,8 +39,7 @@
     "react-native-safe-area-context": "5.4.0",
     "react-native-screens": "~4.11.1",
     "react-native-web": "~0.20.0",
-    "zustand": "^5.0.8",
-    "expo-dev-client": "~5.2.4"
+    "zustand": "^5.0.8"
   },
   "devDependencies": {
     "@babel/core": "^7.25.2",
